IMPALA-8240: Event processor should keep trying when metastore is unavailable.
When metastore is unavailable, EventProcessor state is changed
to ERROR when there is a MetastoreFetchNotificationException.
After this change, the exception handler will not change the state and
EventProcessor continues trying when metastore is unavailable.

Testing:
Added test in MetastoreEventProcessorTest to check event processor
state is active even after multiple NotificationFetchExceptions

Change-Id: I570a52462e6d3b634b2c227dfcb98e20ad2a0023